Checkpoint: start pulling classes out.
