+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/destination_async/StreamDestinationFlusher.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.integrations.destination_async;
+
+import io.airbyte.protocol.models.v0.AirbyteMessage;
+import io.airbyte.protocol.models.v0.StreamDescriptor;
+import java.util.stream.Stream;
+
+public interface StreamDestinationFlusher {
+
+  void flush(StreamDescriptor decs, Stream<AirbyteMessage> stream) throws Exception;
+
+}
