+++ airbyte-integrations/bases/base-java/src/test/java/io/airbyte/integrations/destination_async/BufferManagerEnqueueTest.java
@@ -1,11 +1,16 @@
+/*
+ * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+ */
+
 package io.airbyte.integrations.destination_async;
 
 import org.junit.Test;
 
 public class BufferManagerEnqueueTest {
 
-    @Test
-    void OfferAndRetrieve() {
+  @Test
+  void OfferAndRetrieve() {
 
-    }
+  }
+
 }
